feat: domain, controller, view로 파일 분리
