+++ src/main/kotlin/racingCar/controller/RacingCarController.kt
@@ -1,0 +1,23 @@
+package racingCar.controller
+
+import kotlin.random.Random
+
+class RacingCarController {
+    fun main() {
+        println("자동차 대수는 몇 대인가요?")
+        val numberOfCar = readLine()
+        println("시도할 횟수는 몇 회인가요?")
+        val count = readLine()
+
+        print("$numberOfCar, $count")
+    }
+
+    fun generateRandomNum(gt: Int, lt: Int): Int {
+        return Random.nextInt(gt, lt)
+    }
+
+    fun IsMovingForward(value: Int, ref: Int): Boolean {
+        return value >= ref
+    }
+
+}
