+++ src/main/kotlin/racingCar/domain/Car.kt
@@ -1,4 +1,4 @@
-package racingCar
+package racingCar.domain
 
 class Car {
     val mileage: Int = 0
