+++ src/main/kotlin/racingCar/view/InputView.kt
@@ -1,0 +1,2 @@
+package racingCar.view
+
